typo

+++ src/main/java/ksp/vilius/graph/model/Actor.java
@@ -25,5 +25,5 @@ public class Actor {
     private Date dateOfBirth;
     private String address;
 
-    private Integer filmId;
+    private Long filmId;
 }
